[ERM-4911] fix
git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/ERM.BL/Prerelease/BLFlex;C24338

+++ Source/ApplicationServices/2Gis.Erm.BLFlex.Operations.Global/Shared/Formatter/RussianNumberToWordsConverter.cs
@@ -86,7 +86,7 @@ namespace DoubleGis.Erm.BLFlex.Operations.Global.Shared.Formatter
             }
         }
 
-        public string Convert(int number)
+        public string Convert(long number)
         {
             var minus = false;
             if (number < 0m)
